// src/app/components/models/Home.ts
export class Home {

}

// Imagens do carousel
export class HomeCarousel {
    imagemCarousel: string;
    tituloCarousel: string;
    subtituloCarousel: string;
}
// Enumeradores/Contadores
export class HomeContador {
    enumFichasCriadas: string;
    enumEstoriasCriadas: string;
    enumSessoesAtivas: string;
    enumSorteiosRealizados: string;
}
// O que você encontra na plataforma?
export class HomePlataforma {
    tituloPlataforma: string;
    textoPlataforma: string;
    imagemPlataforma: string;
}
// Estórias Exclusivas da Plataforma
export class HomeEstoriasExclusivas {
    tituloEstoriasExclusivas: string;
    textoEstoriasExclusivas: string;
    imagemEstoriaExclusiva: string;
}
//Sorteio na Plataforma
export class HomeSorteio {
    nomeGanhador: string;
    textoGanhador: string;
    imagemGanhador: string;

    tituloProximoSorteio: string;
    textoProximoSorteio: string;
    imagemProximoSorteio: string;

    tituloHistoricoSorteios: string;
    textoHistoricoSorteios: string;
    imagemHistoricoSorteios: string;
}

// // Variáveis Geral da Page
// export class Home {

//     // Imagens do carousel
//     imgCarousel: string;

//     // Enumeradores/Contadores
//     qntFichasCriadas: number;
//     qntEstoriasCriadas: number;
//     qntSessoesAtivas: number;
//     qntSorteiosRealizados: number;

//     // O que você encontra na plataforma?
//     tituloPlataforma: string;
//     textoPlataforma: string;
//     imagemPlataforma: string;

//     // Estórias Exclusivas da Plataforma
//     tituloEstoriasExclusivas: string;
//     textoEstoriasExclusivas: string;
//     imgEstoriaExclusiva: string;

//     //Sorteio na Plataforma

//     ganhador: string;
//     ultimoGanhador: string;
//     textoGanhador: string;
//     imgGanhador: string;

//     proxSorteio: string;
//     textoProxSorteio: string;
//     imgProxSorteio: string;

//     historicoSorteios: string;
//     textoHistSorteios: string;
//     imgHistSorteios: string;

// }